+++ src/main/java/org/plteco/ploytechcourse/application/user/signup/SendEmailApplication.java
@@ -1,6 +1,6 @@
 package org.plteco.ploytechcourse.application.user.signup;
 
-import org.plteco.ploytechcourse.domain.user.signup.model.dto.emailDto;
+import org.plteco.ploytechcourse.application.user.signup.dto.emailDto;
 
 public interface SendEmailApplication {
     String sendEmail(emailDto emailDto);
