+++ src/main/java/org/plteco/ploytechcourse/application/user/signup/SignupApplication.java
@@ -1,6 +1,6 @@
 package org.plteco.ploytechcourse.application.user.signup;
 
-import org.plteco.ploytechcourse.domain.user.signup.model.dto.SignupUserDto;
+import org.plteco.ploytechcourse.application.user.signup.dto.SignupUserDto;
 
 public interface SignupApplication {
     String signup(SignupUserDto signupUserDto);
