--- src/main/java/org/plteco/ploytechcourse/domain/user/signup/model/dto/emailAndCodeDto.java
@@ -1,9 +1,0 @@
-package org.plteco.ploytechcourse.domain.user.signup.model.dto;
-
-import lombok.Data;
-
-@Data
-public class emailAndCodeDto {
-    private String email;
-    private String code;
-}
